extract version metadata loading in its own hook
